+++ src/main/java/order/controller/service/ShowDeliveryStaffHistoryOrderServlet.java
@@ -28,7 +28,7 @@ public class ShowDeliveryStaffHistoryOrderServlet extends HttpServlet {
         String json = null;
         int userID;
         try {
-            userID = Integer.parseInt(parm);
+            userID = (int)Long.parseLong(parm);
             if(UserDAO.showUserIdentity(userID).equals(MemberSetting.UserStatus.DELIVER_ON)){
                 // 查詢 外送員 歷史訂單
                 OrderDAO.searchDeliverHistoryOrder(userID);
